+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v1.6.0](https://github.com/allenai/beaker-gantry/releases/tag/v1.6.0) - 2024-06-14
+
 ### Added
 
 - Added `-l/--latest`, `-w/--workspace`, `-a/--author` options to `gantry follow` command.